Fix signedness warnings in gio/gwin32appinfo.c
gio/gwin32appinfo.c: In function 'uwp_package_cb':
gio/gwin32appinfo.c:3383:17: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
   for (i = 0; i < supported_extgroups->len; i++)
                 ^
gio/gwin32appinfo.c:3389:29: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
       for (i_ext = 0; i_ext < grp->extensions->len; i_ext++)
                             ^
gio/gwin32appinfo.c:3430:35: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
           for (i_verb = 0; i_verb < grp->verbs->len; i_verb++)
                                   ^
gio/gwin32appinfo.c:3463:33: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
       for (i_hverb = 0; i_hverb < ext->verbs->len; i_hverb++)
                                 ^
gio/gwin32appinfo.c:3478:17: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
   for (i = 0; i < supported_protocols->len; i++)
                 ^
gio/gwin32appinfo.c:3541:33: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
       for (i_hverb = 0; i_hverb < url->verbs->len; i_hverb++)
                                 ^
gio/gwin32appinfo.c: In function 'g_win32_app_info_launch_internal':
gio/gwin32appinfo.c:4799:37: warning: comparison of integer expressions of different signedness: 'gint' {aka 'int'} and 'guint' {aka 'unsigned int'}
           for (p_index = 0; p_index <= g_strv_length (envp); p_index++)
                                     ^~

+++ gio/gwin32appinfo.c
@@ -3359,7 +3359,7 @@ uwp_package_cb (gpointer         user_data,
                 GPtrArray       *supported_extgroups,
                 GPtrArray       *supported_protocols)
 {
-  gint i, i_verb, i_ext;
+  guint i, i_verb, i_ext;
   gint extensions_considered;
   GWin32AppInfoApplication *app;
   gchar *app_user_model_id_u8;
@@ -3459,7 +3459,7 @@ uwp_package_cb (gpointer         user_data,
    */
   while (g_hash_table_iter_next (&iter, NULL, (gpointer *) &ext))
     {
-      gint i_hverb;
+      guint i_hverb;
 
       if (!ext)
         continue;
@@ -3537,7 +3537,7 @@ uwp_package_cb (gpointer         user_data,
 
   while (g_hash_table_iter_next (&iter, NULL, (gpointer *) &url))
     {
-      gint i_hverb;
+      guint i_hverb;
 
       if (!url)
         continue;
@@ -4788,7 +4788,7 @@ g_win32_app_info_launch_internal (GWin32AppInfo      *info,
   if (apppath)
     {
       gchar **p;
-      gint p_index;
+      gsize p_index;
 
       for (p = envp, p_index = 0; p[0]; p++, p_index++)
         if ((p[0][0] == 'p' || p[0][0] == 'P') &&
